Bug fix for put response population
Prior to this fix the key and vector clock from a RpbPutResp were not
being set on the passed in RIACK_OBJECT. This was happening because the
target and src arguments to the RMALLOCCOPY macro were swapped.

Additionally I think this was probably causing a memory leak because the
source len and data pointers were getting overwritten with 0 and thus
the memory that was allocated for them on the heap was no longer
referenced anywhere for freeing.